Issue #3130311 by SV: Create 'Mode Block' ViewsDisplay handler as basic, cleanup sub-module
