--- modules/social_features/social_activity/modules/social_activity_filter/config/install/social_activity_filter.settings.yml
@@ -1,2 +1,0 @@
-vocabulary: {}
-taxonomy_fields: {}
