--- modules/social_features/social_activity/modules/social_activity_filter/config/schema/social_activity_filter.settings.yml
@@ -1,10 +1,0 @@
-social_activity_filter.settings:
-  type: config_object
-  label: 'Social Activity Filter settings'
-  mapping:
-    vocabulary:
-      type: array
-      label: 'Taxonomy vocabularies'
-    taxonomy_fields:
-      type: array
-      label: 'Taxonomy fields'
