--- modules/social_features/social_activity/modules/social_activity_filter/social_activity_filter.info.yml
@@ -1,6 +1,0 @@
-name: 'Social Activity Filter'
-description: 'Provides filtered activity stream for Open Social.'
-type: module
-core: 8.x
-package: Social
-configure: social_activity_filter.settings
