--- modules/social_features/social_activity/modules/social_activity_filter/social_activity_filter.permissions.yml
@@ -1,2 +1,0 @@
-administer social_activity_filter:
-  title: 'Administer Social Activity Filter settings'
