--- modules/social_features/social_activity/modules/social_activity_filter/social_activity_filter.routing.yml
@@ -1,7 +1,0 @@
-social_activity_filter.settings:
-  path: '/admin/config/opensocial/activity-filter-settings'
-  defaults:
-    _form: '\Drupal\social_activity_filter\Form\FilterSettingsForm'
-    _title: 'Social Activity Filter settings'
-  requirements:
-    _permission: 'administer social_activity_filter'
