--- modules/social_features/social_activity/modules/social_activity_filter/social_activity_filter.services.yml
@@ -1,5 +1,0 @@
-services:
-  social_activity_filter.override:
-    class: \Drupal\social_activity_filter\SocialActivityFilterOverride
-    tags:
-      - {name: config.factory.override, priority: 5}
